Set overwrite flag to yes when writing data to XLS
When Slatwall exports a file, it names the file after the date periods,
and it doesn’t delete it after it has been downloaded. So if you go
back and try to export another file of same periods, it returns an
error that the file already exists. This fix will force Slatwall to
overwrite any old file that has same name.